+++ src/main/java/ch/zhaw/schiffeversenken/data/PlayField.java
@@ -3,7 +3,7 @@ package ch.zhaw.schiffeversenken.data;
 import java.util.ArrayList;
 import java.util.List;
 
-import ch.zhaw.schiffeversenken.Coordinate;
+import ch.zhaw.schiffeversenken.helpers.Coordinate;
 
 public class PlayField {
 	private List<Ship> ships;
