+++ src/main/java/ch/zhaw/schiffeversenken/data/Ship.java
@@ -2,7 +2,7 @@ package ch.zhaw.schiffeversenken.data;
 
 import java.util.List;
 
-import ch.zhaw.schiffeversenken.Coordinate;
+import ch.zhaw.schiffeversenken.helpers.Coordinate;
 
 public class Ship {
 	List<Coordinate> shipPositions;
