+++ src/test/java/ch/zhaw/schiffeversenken/gamestart/GameStarterTest.java
@@ -3,11 +3,11 @@ package ch.zhaw.schiffeversenken.gamestart;
 import java.util.ArrayList;
 import java.util.List;
 
-import ch.zhaw.schiffeversenken.Coordinate;
 import ch.zhaw.schiffeversenken.data.Game;
 import ch.zhaw.schiffeversenken.data.PlayField;
 import ch.zhaw.schiffeversenken.data.Ship;
 import ch.zhaw.schiffeversenken.guicomponents.Maingui;
+import ch.zhaw.schiffeversenken.helpers.Coordinate;
 
 public class GameStarterTest {
 	public static void main(String[] args) {
